requirements.txt
pinecone-client==2.2.1
cohere==4.3.0
datasets==2.12.0
redis==4.5.4
numpy==1.21.6
qdrant-client==1.1.6
fastapi==0.94.0
uvicorn==0.21.0
